refactor: seperate contact components
