+++ examples/app-crm/src/components/contact/index.ts
@@ -1,0 +1,2 @@
+export * from "./table-view";
+export * from "./card-view";
